fix: fix tsconfig
