+++ eslintrc.js
@@ -1,0 +1,8 @@
+// .eslintrc.js
+export default {
+    settings: {
+        "import/extensions": [
+            ".js"
+        ]
+    }
+}
