+++ package.json
@@ -2,10 +2,13 @@
   "name": "express-server",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
+    "prebuild": "tslint -c tslint.json -p tsconfig.json --fix",
+    "build": "tsc",
+    "prestart": "npm run build",
+    "start": "node .",
     "create": "node ./services/createTables.js",
-    "start": "nodemon index",
     "start:inspect": "nodemon --inspect index",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -18,14 +21,18 @@
     "express": "^4.18.2",
     "make-runnable": "^1.4.1",
     "mtgsdk": "^1.0.1",
-    "pg": "^8.11.3",
-    "typescript": "^5.3.3"
+    "pg": "^8.11.3"
   },
   "devDependencies": {
     "@types/body-parser": "^1.19.5",
     "@types/express": "^4.17.21",
+    "@types/node": "^16.18.80",
     "@types/pg": "^8.11.0",
-    "nodemon": "^3.0.3"
+    "@typescript-eslint/parser": "^6.21.0",
+    "eslint-plugin-import": "^2.29.1",
+    "nodemon": "^3.0.3",
+    "tslint": "^6.1.3",
+    "typescript": "^4.9.5"
   },
   "type": "module"
 }
