+++ src/cloudinary.ts
@@ -2,7 +2,7 @@ import {} from 'dotenv/config'
 import {v2 as cloudinary} from 'cloudinary';
 
 export const initCloudinary = () => {
-    cloudinary.config({ 
+    cloudinary.config({
         cloud_name: process.env.CLOUD_NAME,
         api_key: process.env.API_KEY,
         api_secret: process.env.API_SECRET
